community-solutions/ai-chat/protegrity-ai-llm/app/requirements.txt
asgiref==3.11.0
boto3==1.35.95
certifi==2024.8.30
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
iniconfig==2.3.0
openai==1.57.4
PyJWT==2.10.1
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.0.0
requests==2.32.3
sqlparse==0.5.3
